+++ crates/config/src/toolchain/mod.rs
@@ -1,6 +1,7 @@
 mod bin_config;
 mod bun_config;
 mod deno_config;
+mod moon_config;
 mod node_config;
 mod python_config;
 mod rust_config;
@@ -9,6 +10,7 @@ mod typescript_config;
 pub use bin_config::*;
 pub use bun_config::*;
 pub use deno_config::*;
+pub use moon_config::*;
 pub use node_config::*;
 pub use python_config::*;
 pub use rust_config::*;