+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 1.29.1
+
+#### 🚀 Updates
+
+- Added a new task option `merge`, that defines the same strategy for all applicable fields. Can be
+  overridden with `mergeX` task options.
+- Added a `moon` setting to `.moon/toolchain.yml`, which can be used to customize the update check
+  process.
+
+#### 🐞 Fixes
+
+- Fixed `moon query projects` including dependency projects by default. Can be controlled with
+  `--upstream`.
+- Fixed `moon query projects` regex patterns not working when used in conjunction with affected.
+- Fixed Bash-based hooks being generated with the wrong path separators on Windows.
+- Fixed an issue where an inherited task with merge strategy "replace" will accidentally remove task
+  args, deps, and env in inheriting tasks.
+- Fixed an issue where token variables and environment variables used within task args together
+  would not be replaced correctly.
+
 ## 1.29.0
 
 #### 🚀 Updates